Add venue filter and first test
